+++ package.json
@@ -8,9 +8,10 @@
     "@formspree/react": "^2.5.0",
     "immer": "^10.0.2",
     "luxon": "^3.4.0",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.6.3",
     "react-scripts": "5.0.0",
-    "react": "^17.0.2",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
   },
@@ -21,8 +22,8 @@
     "@types/jest": "^27.4.1",
     "@types/luxon": "^3.3.1",
     "@types/node": "^16.11.26",
-    "@types/react-dom": "^17.0.14",
-    "@types/react": "^17.0.41"
+    "@types/react": "^17.0.41",
+    "@types/react-dom": "^17.0.14"
   },
   "scripts": {
     "start": "react-scripts start",
